move file to the right structure
